WRP-6648:DayPicker: fix TypeError when "selected" is a number
DayPicker accepts a number or an array of numbers in "selected" prop.
But TypeError exception occurs in localizedSelected() function when passing
a number (not 0).

localizedSelected() function also accepts a number or an array of numbers
as a "selected" argument.
The problem is this.
 1. the function doesn't differentiate between number 0 and undefined/null.
 2. treats unconditionally the "selected" argument as an array.

The fix is this.
 1. differentiate between 0 and undefined/null
 2. check type of "selected" argument and if it's a number, then returns a
    localized number.
 3. leave the selected.map() for arrays as a fallback.

Enact-DCO-1.0-Signed-off-by: Hoeun Ryu <hoeun.ryu@lge.com>