WRP-8078:Button: Add roundBorder prop making both sides of button fully round
As a result of a discussion with the GUI team, it was decided to provide a
boolean prop to have both the left and right sides of the button completely
round. Therefore, a `roundBorder` boolean prop is added to achieve that.

Use `border-radius: 9999px`.

* https://w3c.github.io/csswg-drafts/css-backgrounds/#corner-overlap

Corner curves must not overlap: When the sum of any two adjacent
border radii exceeds the size of the border box, UAs must
proportionally reduce the used values of all border radii until none
of them overlap.

You can also use `border-radius: calc($height / 2)` for the exact
height, but that will require a .sand-custom-text mixin nested inside
applySkins mixin for the "large text" environment.

    .sand-custom-text({
        &.roundBorder {
	    .bg {
	         border-radius: calc(@sand-button-height-large / 2);
            }
        }
    });

To avoid this, we use the property of CSS specification that corner
curves do not overlap when the border-radius is set to a tremendous
value.

Enact-DCO-1.0-Signed-off-by: Hoeun Ryu <hoeun.ryu@lge.com>