WRP-8078:Button: Add roundBorder prop making both sides of button fully round
As a result of a discussion with the GUI team, it was decided to provide a
boolean prop to have both the left and right sides of the button completely
round. Therefore, a `roundBorder` boolean prop is added to achieve that.

Use
- `--button-height: $height`
- `border-radius: calc(--button-height / 2)`.

Use CSS custom property `--button-height` and LESS property as
variable `$height`. CSS custom property `--button-height` is
overridden wherever CSS property `height` is defined.

Enact-DCO-1.0-Signed-off-by: Hoeun Ryu <hoeun.ryu@lge.com>