WRP-10826:WizardPanels: add 40ms delay to setTimeout in onTranstion
Until 2.5.8, focus exists always on the screen in 5-way mode when
next/prev button of WizardPanels sample is selected. From 2.6.0, focus
disappears when moving from view 3 to view 4 via next button and
moving from view 2 to view 1 via prev button.

This behavior is observed on PC, and also ui-test on local machine is
possibly failed due to this. The failed test case is described as
"should select buttons over header - [QWTC-2390]" in WizardPanels
specs. Note that we don't have this issue on TV and jenkins tests yet.

This problem is observed in Chrome version 104 or higher, but not in
Chrome version 103 or lower.

The observed problems are as follows and this means Spotlight refocus
on the item that was previously focused on in the previous view.
- When transitioning from View 3 to View 4, the focus still remains on
  Item (Hello Item) in View 3.
- When transitioning from View 2 to View 1, the focus still remains on
  Button A in View 2.

The problem goes away if adding some delay in `setTimeout()` in
`onTransition` Handler. There seems to be race condition between view
transition and rendering.

The delay of "40ms" is to ensure that the timer handler is called
after two rendering cycles (60HZ).

Enact-1.0-DCO-Signed-off-by: Hoeun Ryu <hoeun.ryu@lge.com>