WRP-6057: Stop scrolling by hoverToScroll when the pointer disappears
 While scrolling by "hoverToScroll", scrolling should stop when the
pointer disappears due to timeout, but it does not stop. This
phenomenon is reproduced on TV images (not RP).

* Difference between RP and TV
- In RP, the `getLastInputType` luna call always fails, so
  `input.activated` remains `false`. Then `setInputType('key')` is
  always called in `SpotlightRootDecorator::handleKeyDown` so the RAF
  handler can detect this and call `stopRaf`.
- In TV, `getLastInputType` luna call succeeds and `input.activated`
  remains `true`. Therefore, since `setInputType('key')` is not called
  in `SpotlightRootDecorator::handleKeyDown`, the RAF handler does not
  detect the input type change and does not get a chance to call
  `stopRaf`.
- `getLastInputType` only exists in tv API

* Causes
 The root cause is that `com.webos.surfacemanager` does not give
`getLastInputType` LS message when the pointer disappears after a
certain period of time and changes to key input mode.
 Instead, a key code is generated as if a key was pressed, and
`SpotlightRootDecorator::handleKeyDown` is called. When the pointer
disappears, key code 1537 occurs, and when the pointer appears, key
code 1536 occurs.
 If an LS message is given whenever the input type changes,
`setInputType` can be called accordingly to change the input type,
and `stopRaf` can be called when the RAF handler detects this.
 However, this API has different usability for each platform, so this
API cannot depend on the input type change mechanism.
 Seeing this limitation as a behavioral characteristic of
`surfacemanager` and if it can't be fixed, we need to think of another
 way.

* Suggestions
 A flag `stopHoverByScroll` should be set (true) in
`HoverToScroll::handleGlabalKeyDown` if the keycode is `pointerHide`.
`scrollByHover` callback of `requestAnimationFrame` exits the callback
loop if the flag is set.

Enact-DCO-1.0-Signed-off-by: Hoeun Ryu <hoeun.ryu@lge.com>