WRP-13074: call getLastInputType method only when it's available
This issue was raised because we unconditionally execute
`getLastInputType` LS call without checking for the existence of the
function.

The `getLastInputType` method only exists on TV platforms, not RP
or OSE. So we need to check the function's existence first before
calling the function.

We can query the API method list using a special method.
- `com/palm/luna/private/introspection`
It returns a JSON object, and the method list can be extracted using
'/' key. So, you write code like this.

```javascript
    new LS2Request().send({
        service: 'luna://com.webos.surfacemanager',
        method: 'com/palm/luna/private/introspection',
        onSuccess: function ({'/': apiList}) {
            console.log(apiList);
        }
    });
```

Using this API, ThemeDecorator checks the existence of
`getLastInputType` function in `constructor` and calls the function
only when it is available in `componentDidMount`

Enact-DCO-1.0-Signed-off-by: Hoeun Ryu <hoeun.ryu@lge.com>