WRP-14562:WizardPanels: fix audio guidance for step when using `current`
Since `aria-label` only uses the `index` value, `current` is not used
when `current` is passed as a prop.
The fix is that if `current` is passed as a prop, this value is used
to set `aria-label`. This patch includes a unit test that checks that
`aria-label` is set with the `current` value when `current` is passed
as a prop.

Enact-DCO-1.0-Signed-off-by: Hoeun Ryu <hoeun.ryu@lge.com>