WRR-4590: Fix to focus a node properly after scroll by voice control
When a focusable node inside a scroller is sticked to the top,
it cannot be focused after scroll by voice control.

The cause is a logic to check if the candidate focus target is
inside of a scroller viewport; missing the case that the boundaries
of a viewport and a candidate are equal.

Enact-DCO-1.0-Signed-off-by: Seungcheon Baek (sc.baek@lge.com)